makingUI: ForEach를 이용해 ButtonView 리팩토링
